server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.24.4
python-dotenv==1.0.0
requests==2.31.0
